cycle_outlook/requirements.txt
Flask==3.0.2
pywin32==306
requests==2.32.3
Jinja2==3.1.4
PyYAML==6.0.1
pytest==8.2.0
